services/lambda_data_fetcher/pyproject.toml
[project]
name = "lambda_data_fetcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "connectorx>=0.4.2",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "yfinance>=0.2.54",
]

[dependency-groups]
dev = [
    "fastexcel>=0.13.0",
    "jupyter>=1.1.1",
    "mypy>=1.15.0",
    "ruff>=0.9.7",
    "types-requests>=2.32.0.20241016",
]

[tool.mypy]
# 基本設定
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

# third-party ライブラリの型情報を使う
ignore_missing_imports = true

[tool.ruff]
exclude = [
    ".venv",
    ".ruff_cache",
]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    'F', # Flake8
    'B', # Black
    'I', # isort
    'E', # Error
    'W', # Warning
]
extend-select = ["I"]
ignore = []
fixable = []
unfixable = []
pydocstyle.convention = 'google'

[tool.ruff.lint.pylint]
max-args = 6

[tool.uv.workspace]
members = ["lambda_data_fetcher"]
